Function to scan an HBase table and return a data frame.
The row names of the resulting data frame are the keys of the HBase
table. The column names are those specified in colspec.  If a row does
not have a value for a column specified in colspec, then that cell will
be populated with NA.

Prep work to allow all columns in a column family.

Use cols instead of colspec when getting column values.

modularise helper methods